[#8] 마이페이지 세부페이지의 GNB 제거 및 CustomSubAppBar 생성 및 적용
#8
