+++ lib/core/router/route_path.dart
@@ -4,7 +4,7 @@ class RoutePath {
 
   static const String myPage = '/my_page/my_page_screen';
   static const String myPage_notice = '/my_page/notice/notice_screen';
-  static const String myPage_subscription_payment_info = '/my_page/subscription_payment_info/subscription_payment_info_screen';
+  static const String myPage_album_payment_info = '/my_page/album_payment_info/album_payment_info_screen';
   static const String myPage_address_management = '/my_page/address_management/address_management_screen';
   static const String myPage_add_address = '/my_page/address_management/add_address_screen';
   static const String myPage_subscription_history = '/my_page/album_subscription_history/album_subscription_history_screen';
