+++ lib/presentation/screens/my_page/settings/setting_screen.dart
@@ -3,7 +3,7 @@ import 'package:cherrypic/core/constants/font.dart';
 import 'package:cherrypic/presentation/screens/my_page/settings/setting_toggle.dart';
 import 'package:flutter/material.dart';
 
-import '../../../widgets/custom_tab_bar.dart';
+import '../../../widgets/custom_sub_app_bar.dart';
 import '../common_popup_dialog.dart';
 
 class SettingScreen extends StatefulWidget {
@@ -45,10 +45,10 @@ class _SettingScreenState extends State<SettingScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      appBar: const CustomSubAppBar(title: '설정'),
       backgroundColor: Colors.white,
       body: Column(
         children: [
-          const CustomTabBar(title: '설정'),
           const SizedBox(height: 29.28),
           Padding(
             padding: const EdgeInsets.symmetric(horizontal: 30),
